# Some runtime optimizations