--- a/src/abi/returning.rs
+++ b/src/abi/returning.rs
@@ -58,7 +58,7 @@ pub fn codegen_with_call_return_arg<'tcx, B: Backend, T>(
     let return_ptr = match output_pass_mode {
         PassMode::NoPass => None,
         PassMode::ByRef => match ret_place {
-            Some(ret_place) => Some(ret_place.to_addr(fx)),
+            Some(ret_place) => Some(ret_place.to_ptr(fx).get_addr(fx)),
             None => Some(fx.bcx.ins().iconst(fx.pointer_type, 43)),
         },
         PassMode::ByVal(_) | PassMode::ByValPair(_, _) => None,