@bjorn3 2019-12-20
The last commit uses `stack_load` and `stack_store` where possible. While Cranelift currently expands them to `stack_addr` + `load`/`store`, it can give a performance improvement once Cranelift uses a single `mov` on x86.